Document cryptocontext parameters (#222)
* Corrected comments and changed the type of maxRelinSkDegi from int to uint32_t

* Updated README.md with cryptoparameter descriptions and instructions how to build user applications

* Add missing examples and fix newlines.

* Updated file names and add link to default values.

+++ src/pke/include/schemebase/base-cryptoparameters.h
@@ -129,7 +129,7 @@ public:
    *
    * @return maximum power of secret key
    */
-    virtual size_t GetMaxRelinSkDeg() const {
+    virtual uint32_t GetMaxRelinSkDeg() const {
         return 0;
     }
